changes amount from double to numeric

+++ db/create-table.sql
@@ -7,7 +7,7 @@ CREATE TABLE IF NOT EXISTS public.tb_payments
     promo character varying(255) COLLATE pg_catalog."default",
     status integer NOT NULL,
     "customerAddress" character varying(255) COLLATE pg_catalog."default",
-    amount double precision NOT NULL,
+    amount numeric NOT NULL,
     currency character varying(255) COLLATE pg_catalog."default" NOT NULL,
     "startTS" bigint NOT NULL,
     "endTS" bigint NOT NULL,
